(WS23PASTST-28): Update Backend-Server and base implementation logic

+++ backend/src/main/java/com/thws/ar/stolpersteine/backend/db/entity/Victim.java
@@ -12,7 +12,7 @@ public class Victim {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer victim_id;
+    private Long victim_id;
 
     @Column(name = "name")
     private String name;
