+++ backend/src/main/java/com/thws/ar/stolpersteine/backend/db/repositories/StolpersteinRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StolpersteinRepository extends JpaRepository<Stolperstein, Integer> {
+public interface StolpersteinRepository extends JpaRepository<Stolperstein, Long> {
 }
